requirements.txt
bleach==1.4.3
codeclimate-test-reporter==0.1.2
coverage==4.1
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.9.7
django-filter==0.13.0
django-nose==1.4.3
djangorestframework==3.4.0
gunicorn==19.6.0
html5lib==0.9999999
Markdown==2.6.6
nose==1.3.7
oauthlib==1.1.2
psycopg2==2.6.1
PyJWT==1.4.1
python-social-auth==0.2.19
python3-openid==3.0.10
pytz==2016.6.1
requests==2.10.0
requests-oauthlib==0.6.2
six==1.10.0
whitenoise==3.2
